apps/api/requirements.txt
# Core FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode==7.4.2

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# Cryptography
cryptography==42.0.0

# WebSockets
websockets==12.0

# Background Tasks
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.2.0

# Code Quality
ruff==0.1.14
mypy==1.8.0
types-python-jose==3.3.4.8
types-passlib==1.7.7.20240106

# Market Data & Trading (Adapters)
alpaca-py==0.16.0
pandas==2.1.4
numpy==1.26.3

# Plaid (Bank Linking)
plaid-python==18.0.0
